feat: edit username

+++ src/functions/user/login.ts
@@ -8,6 +8,7 @@ import io from '../../utils/io-util.js'
 // FUNCTIONS
 import { createNewUser } from './create-new-user.js'
 import { deleteUser } from './delete-user.js'
+import { editUser } from './edit-user.js'
 
 // TYPES
 import type { User } from '../../shared/types/user.js'
@@ -27,16 +28,21 @@ export async function login(): Promise<User> {
       choices: [
         { name: '+ Create new user', value: 'create-new-user' },
         { name: '- Delete user', value: 'delete-user' },
+        { name: '# Edit user', value: 'edit-user' },
         ...users.map(user => ({ name: user.username, value: user.id }))
       ]
     })
   
-    switch (userSelection) {
-      case 'create-new-user': return await createNewUser()
-      // case 'edit-user': return await editUser()
-      case 'delete-user':
-        await deleteUser()
-        return await login()
+    if (['create-new-user', 'delete-user', 'edit-user'].includes(userSelection)) {
+      switch (userSelection) {
+        case 'create-new-user': return await createNewUser()
+        case 'edit-user':
+          await editUser()
+          return await login()
+        case 'delete-user':
+          await deleteUser()
+          return await login()
+      }
     }
   
     const user = users.find(u => u.id === userSelection)
